requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
defusedxml==0.7.1
discord.py==2.1.0
et-xmlfile==1.1.0
fonttools==4.46.0
fpdf2==2.7.6
frozenlist==1.3.3
idna==3.4
lxml==4.9.3
multidict==6.0.3
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==10.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
requests==2.28.1
secure-smtplib==0.1.1
six==1.16.0
slack-sdk==3.26.1
slackclient==2.9.4
soupsieve==2.5
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2

